requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
deep-translator==1.11.4
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
Pillow==10.0.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.3
wikipedia==1.4.0
yarl==1.9.2
